# swagger/swagger-api.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/beneficiarios":{"get":{"tags":["beneficiario-controller"],"operationId":"listar","parameters":[{"name":"paginacao","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageBeneficiarioResponse"}}}}},"security":[{"bearer-key":[]}]},"put":{"tags":["beneficiario-controller"],"operationId":"atualizar","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BeneficiarioRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BeneficiarioResponse"}}}}},"security":[{"bearer-key":[]}]},"post":{"tags":["beneficiario-controller"],"operationId":"cadastrar","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BeneficiarioPostRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BeneficiarioResponse"}}}}},"security":[{"bearer-key":[]}]}},"/login":{"post":{"tags":["autenticacao-controller"],"operationId":"efetuarLogin","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DadosAutenticacao"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DadosTokenJWT"}}}}}}},"/beneficiarios/{id}/documentos":{"get":{"tags":["beneficiario-controller"],"operationId":"listarDocumentosPorIdBeneficiario","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DocumentoResponse"}}}}}},"security":[{"bearer-key":[]}]}},"/beneficiarios/{id}":{"delete":{"tags":["beneficiario-controller"],"operationId":"excluir","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BeneficiarioResponse"}}}}},"security":[{"bearer-key":[]}]}}},"components":{"schemas":{"BeneficiarioRequest":{"required":["id"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nome":{"type":"string"},"telefone":{"type":"string"},"dataNascimento":{"type":"string","format":"date"}}},"BeneficiarioResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nome":{"type":"string"},"email":{"type":"string"},"dataNascimento":{"type":"string","format":"date"},"dataInclusao":{"type":"string","format":"date-time"}}},"DadosAutenticacao":{"type":"object","properties":{"login":{"type":"string"},"senha":{"type":"string"}}},"DadosTokenJWT":{"type":"object","properties":{"token":{"type":"string"}}},"BeneficiarioPostRequest":{"required":["dataNascimento","nome","telefone"],"type":"object","properties":{"nome":{"type":"string"},"telefone":{"type":"string"},"dataNascimento":{"type":"string","format":"date"},"documentos":{"type":"array","items":{"$ref":"#/components/schemas/DocumentoPostRequest"}}}},"DocumentoPostRequest":{"required":["descricao","tipoDocumento"],"type":"object","properties":{"tipoDocumento":{"type":"string","enum":["RG","CERTIDAO_NASCIMENTO","CNH","PASSAPORTE"]},"descricao":{"type":"string"}}},"Pageable":{"type":"object","properties":{"page":{"minimum":0,"type":"integer","format":"int32"},"size":{"minimum":1,"type":"integer","format":"int32"},"sort":{"type":"array","items":{"type":"string"}}}},"PageBeneficiarioResponse":{"type":"object","properties":{"totalPages":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/BeneficiarioResponse"}},"number":{"type":"integer","format":"int32"},"sort":{"$ref":"#/components/schemas/SortObject"},"first":{"type":"boolean"},"last":{"type":"boolean"},"numberOfElements":{"type":"integer","format":"int32"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"empty":{"type":"boolean"}}},"PageableObject":{"type":"object","properties":{"offset":{"type":"integer","format":"int64"},"sort":{"$ref":"#/components/schemas/SortObject"},"unpaged":{"type":"boolean"},"paged":{"type":"boolean"},"pageNumber":{"type":"integer","format":"int32"},"pageSize":{"type":"integer","format":"int32"}}},"SortObject":{"type":"object","properties":{"empty":{"type":"boolean"},"unsorted":{"type":"boolean"},"sorted":{"type":"boolean"}}},"DocumentoResponse":{"required":["id"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"tipoDocumento":{"type":"string","enum":["RG","CERTIDAO_NASCIMENTO","CNH","PASSAPORTE"]},"descricao":{"type":"string"}}}},"securitySchemes":{"bearer-key":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}}}